correções

+++ src/Contract/Service/PatrimonyContract/IGenericService.cs
@@ -9,12 +9,12 @@ namespace SGP.Contract.Service.PatrimonyContract
         Task<List<T>> GetAll();
         Task<T> Details(long? id);
         Task<T> Create(T obj);
-        Task<T> GetUpdate(int id);
+        Task<T> GetUpdate(long id);
         Task<T> Update(long id, T obj);
-        Task<T> Delete(int? id);
-        Task<T> DeleteConfirmed(int id);
+        Task<T> Delete(long? id);
+        Task<T> DeleteConfirmed(long id);
         Task<bool> Exists(long id);
 
-
+        //Task<TReturn> Create<TReturn>(T obj);
     }
 }
